Postgresql database (#3)
* Replacing sqlite3 to Postgresql

This is a major update, be aware of bugs! Tested but not enough to be done.

* Improving stability

Fixing bugs and improving bugs related to the Postgresql database.

* Fix many bugs and improve MDs and code

Fixed many bugs that could (and some happened) that the value of PQgetvalue could corrupt itself.
Also improved the MDs to fix some warns.
Added config.json postgresql field.
Optimizations related to postgresql made (don't make unecessary requests to the database if not really using).

* Fixing memory leaks and improving Dockerfile

In some cases a event goes to a specific if, it doesn't execute PQfinish to close the database connection, leading to many memory leaks. Dockerfile improved with now working properly.

* Remove Dockerfile and add build.yml

Removing Dockerfile since there is no longer Docker based workflows and replace it with build.yml.